bart/pyproject.toml
[tool.poetry]
name = "bart"
version = "1.0"
description = ""
authors = ["Jacob Waksmacki <jwaksmac@med.umich.edu>"]

[tool.poetry.scripts]
art = "bart.__main__:main"

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^0.19.2"
globus-sdk = "^3.2.1"
joblib = "^1.1.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
